+++ src/Profile.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace LaravelGravatar;
+
+use Gravatar\Profile as BaseProfile;
+
+class Profile extends BaseProfile
+{
+}